renamed createlogger to logger
